Move ParserFormatSAT to Internal namespace
