+++ src/Internal/ParserFormatSAT.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PhpCfdi\CfdiSatScraper;
+namespace PhpCfdi\CfdiSatScraper\Internal;
 
 /**
  * This class is a helper to parse that incredibly weird responses from SAT that contains internal
@@ -11,6 +11,8 @@ namespace PhpCfdi\CfdiSatScraper;
  * format: |value-length|field-type|field-name|value
  * example: |3|hiddenField|__FOO|foo|0|hiddenField|__EMPTY||16|hiddenField|__BAR|0123456789abcdef|
  * contents: __FOO: foo, __EMPTY: , __BAR: 0123456789abcdef
+ *
+ * @internal
  */
 class ParserFormatSAT
 {
